More Windows-specific fixes

+++ analyzer/impl/multicast.h
@@ -44,8 +44,8 @@
 # define su_htonll(x) (x)
 # define su_ntohll(x) (x)
 #else
-# define su_htonll(x) (((uint64_t)htonl((x) & 0xFFFFFFFF) << 32) | htonl((x) >> 32))
-# define su_ntohll(x) (((uint64_t)ntohl((x) & 0xFFFFFFFF) << 32) | ntohl((x) >> 32))
+# define su_htonll(x) (((uint64_t)htonl((x) & 0xFFFFFFFF) << 32) | htonl((uint64_t) (x) >> 32))
+# define su_ntohll(x) (((uint64_t)ntohl((x) & 0xFFFFFFFF) << 32) | ntohl((uint64_t) (x) >> 32))
 #endif
 
 /************************** Multicast manager *******************************/
